requirements.txt
boto3>=1.34
pandas>=2.0
jinja2>=3.1
click>=8.1
python-dotenv>=1.0
msal>=1.28
requests>=2.31
psycopg2-binary>=2.9
openpyxl>=3.1
